Version 2 of Nager.Date (#601)
Project structure changed, changes in the DateSystem this has been split into two parts, there is now a WeekendSystem and a HolidaySystem. Class renaming,...

The `Date` of a holiday now returns the original date of the holiday if the date on which it is celebrated is needed there is now an additional property `ObservedDate`.