chore: Extend Dependabot Support
The PR extends the functionality beyond the EcoSystem `nuget`, so that we will also receive dependency updates for `GitHub-Actions` in the future.

In addition, the PR build pipeline has been extended so that an additional step is now performed in the case of `dependabot[bot]` pull requests, which automates these PRs `merged` based on the configurations.